requirements.txt
# 뇌출혈 진단 시스템 - 필수 패키지

# Deep Learning
tensorflow>=2.10.0
keras>=2.10.0

# Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0
pydicom>=2.4.0  # DICOM 지원 (선택적)

# Web Framework
streamlit>=1.28.0

# Data Science
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
scikit-learn>=1.3.0

# Visualization
seaborn>=0.12.0

# Utilities
tqdm>=4.66.0
